requirements.txt
torch
numpy
pandas
matplotlib
joblib
pillow